requirements.txt
Flask
Flask-PyMongo
pymongo
python-dotenv
gunicorn
reportlab
